rename log-api.ts -> collection-log-api.ts move collection-log-api.ts and responses.ts to api/collection-log-api
